pike-rag-gui/backend/services/document_service.py
from typing import List, Optional
from pathlib import Path
import uuid
from datetime import datetime
import magic
import asyncio
from sqlalchemy.ext.asyncio import AsyncSession

from ..models import Document, FileType, DocumentResponse
from ..database import SessionLocal

class DocumentService:
    def __init__(self):
        self.db = SessionLocal()

    async def process_document(self, file_path: Path) -> Document:
        """处理上传的文档"""
        # 检测文件类型
        mime = magic.Magic(mime=True)
        file_type = mime.from_file(str(file_path))
        
        # 根据MIME类型确定文件类型
        if file_type == "application/pdf":
            doc_type = FileType.PDF
        elif file_type == "text/plain":
            doc_type = FileType.TXT
        elif file_type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
            doc_type = FileType.DOCX
        elif file_type == "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet":
            doc_type = FileType.XLSX
        elif file_type == "text/csv":
            doc_type = FileType.CSV
        elif file_type == "text/markdown":
            doc_type = FileType.MD
        else:
            raise ValueError(f"Unsupported file type: {file_type}")

        # 创建文档记录
        doc = Document(
            id=str(uuid.uuid4()),
            filename=file_path.name,
            file_type=doc_type,
            created_at=datetime.now(),
            processed_at=datetime.now()
        )

        # TODO: 调用PIKE-RAG的文档处理逻辑
        # 这里需要集成PIKE-RAG的文档处理功能

        return doc

    async def get_all_documents(self) -> List[DocumentResponse]:
        """获取所有文档列表"""
        # TODO: 从数据库获取文档列表
        return []

    async def get_document_preview(self, doc_id: str) -> str:
        """获取文档预览内容"""
        # TODO: 从数据库获取文档预览
        return ""

document_service = DocumentService() 